+++ provision/vpopmail.sh
@@ -9,7 +9,7 @@ export JAIL_CONF_EXTRA=""
 export JAIL_FSTAB="$ZFS_DATA_MNT/vpopmail/home $ZFS_JAIL_MNT/vpopmail/usr/local/vpopmail nullfs rw 0 0"
 
 export VPOPMAIL_OPTIONS_SET=""
-export VPOPMAIL_OPTIONS_UNSET="ROAMING"
+export VPOPMAIL_OPTIONS_UNSET="ROAMING PGSQL LDAP ORACLE SYBASE"
 
 mt6-include vpopmail
 mt6-include mysql
@@ -148,6 +148,11 @@ mail_qmailadmin_UNSET=CATCHALL CRACKLIB DOMAIN_AUTOFILL IDX_SQL SPAM_DETECTION S
 
 	export WEBDATADIR=../../data/htdocs CGIBINDIR=../../data/cgi-bin CGIBINSUBDIR=qmailadmin
 
+	# port requires newer autoconf than pkg has, 2026-04-04
+	stage_make_conf devel_autoconf_ "
+devel_autoconf_UNSET=INFO
+"
+	stage_port_install devel/autoconf
 	stage_port_install mail/qmailadmin
 
 	install_lighttpd
@@ -416,7 +421,7 @@ migrate_vpopmail_home()
 	# fi
 
 	# # TODO: patch fstab mounts in /etc/jail.conf
-	# service jail stop dovecot vpopmail
+	# service jail start dovecot vpopmail
 }
 
 migrate_vpopmail_home